Removed prefix use in setting SCRIPT_NAME
Why you made the change:

I did this because the prefix is no longer needed for setting
SCRIPT_NAME as setting of the SCRIPT_NAME and PATH_INFO rack environment
variables has been corrected via the overwrite of the
`rewrite_partial_path_info` method. The `rewrite_path_info` method looks
to already be implemented correctly in http_router proper.

I also cleaned up the code by eliminating the 'PATH_INFO' and
'SCRIPT_NAME' magic strings and replacing them with the appropriate
constants. Beyond that, I improved the request_url_test so that we could
properly verify both the path and url helpers as well as the setting of
the SCRIPT_NAME and PATH_INFO.