Make hanami routes work with nested non-hanami apps
Why you made the change:

I did this so that non-hanami applications like Sinatra, apps, etc. when
mounted will still allow the `hanami routes` command line interface to
work.

How the change addresses the need:

As discussed in issue #92 there is an issue where when a non-hanami
application is mounted that responds to the `routes` method that doesn't
return a `Hanami::Router` it throws an exception. This change fixes that
by identifying non-hanami nested routers properly rather than just doing
it based the application responding to the `routes` method.