Make nested_router use duck-type checking
Why you made the change:

I did this because @jodosha recommended it as a more flexible path. I
poked around the code to make sure that there were no further behavioral
checks necessary at this point.

I also update the documentation to better communicate the intent of the
method rather than the previous documentation which communicated the
intent as explicitly identifying a nested Hanami::Router.