Skip unknown content types when parsing body, expand parser loading
When a BodyParser middleware is loaded, but the request’s content type does not match any of the loaded parser, do not attempt to parse and symbolize the body. This ensures there will be no unexpected content in the resulting request env hash.

Also, expand parser loading to support already-initialized Parser instances.