Allow the version of Rack to determine whether a content-length header is present
Hanami Router is meant to be a layer as thin as possible over Rack. It’s not really the router’s job to dictate whether a content-length header should be in the response. That’s up to your app or your middleware.

Content-length would appear in all responses as a side effect of a Rack 2 implementation detail. We don’t need to preserve that as we add Rack 3 support.